Make sure there is single instance of BuildsRepository around.
That limitation comes from the fact that BuildsRepository holds simple
data structure for storing builds in memory and it must be shared by
worker threads.

While this data structure may be separated in the future, allowing the
existence of many BuildsRepository instances I don't see a strong reason
to do it now.